requirements.txt
requests==2.31.0
PyGithub==2.1.1
GitPython==3.1.40
Jinja2==3.1.2
click==8.1.7
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
pylint==3.0.3
pyyaml==6.0.1
